requirements.txt
streamlit
python-dotenv
youtube-transcript-api
langchain
langchain-groq
langgraph
langchain-text-splitters